+++ daemon/libnetwork/drivers/bridge/internal/iptabler/iptabler.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/libnetwork/drivers/bridge/internal/firewaller"
+	"github.com/docker/docker/daemon/libnetwork/internal/modprobe"
 	"github.com/docker/docker/daemon/libnetwork/iptables"
-	"github.com/docker/docker/internal/modprobe"
 )
 
 const (
