+++ daemon/libnetwork/ns/init_linux.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/containerd/log"
-	"github.com/docker/docker/internal/modprobe"
+	"github.com/docker/docker/daemon/libnetwork/internal/modprobe"
 	"github.com/docker/docker/internal/nlwrap"
 	"github.com/vishvananda/netns"
 )
